+++ bazel_tidy_example/return_pointer_full.cpp
@@ -1,4 +1,32 @@
 #include <iostream>
+class Lane{};
+
+Lane* GetMapLaneById(const int Id){
+  if(Id<0){
+    return nullptr;
+  }else{        
+    return new Lane;
+  }
+}
+
+void MapFusionConsumer(){
+
+  const auto lane_ptr_maybe = GetMapLaneById(true);
+
+  if (!lane_ptr_maybe){
+    std::cout << " Lane Creation Failed\n";
+  }
+
+  
+  const auto another_lane_ptr_maybe = GetMapLaneById(false);
+  if (!another_lane_ptr_maybe){
+    std::cout << " Lane Creation Failed\n";
+  }
+
+  // DONE Checking, do rest of the work
+
+}
+
 class A{};
 
 A* UpperStreamFunction(bool flag){
@@ -31,3 +59,6 @@ void DownstreamConsumer(){
   std::cout<<" Happy Case\n";
 }
 
+
+
+// What is not covered